+++ include/fastlog.hpp
@@ -12,17 +12,17 @@
 
 #include "args.hpp"
 
-extern std::vector<float> fastlog_cache;
+extern std::vector<double> fastlog_cache;
 
 /// Initializes the cache with logs of small values.
 void init_fastlog(size_t x);
 
-inline float fastlog(size_t x) {
+inline double fastlog(size_t x) {
     if (fastlog_cache.size() < args.cachesize) {
         init_fastlog(args.cachesize);
     }
     if (x >= fastlog_cache.size()) {
-        return logf(float(x));
+        return logf(double(x));
     }
     return fastlog_cache[x];
 }
