+++ include/sample.hpp
@@ -12,33 +12,33 @@
 namespace sample {
 
 struct ES_State {
-    ES_State(int num_vertices) {
-        this->contribution = utils::constant<int>(num_vertices, 0);
+    ES_State(long num_vertices) {
+        this->contribution = utils::constant<long>(num_vertices, 0);
         this->neighborhood_flag = utils::constant<bool>(num_vertices, false);
-        this->neighbors = std::set<int>();
+        this->neighbors = std::set<long>();
         this->contribution_sum = 0;
     }
-    std::vector<int> contribution;
+    std::vector<long> contribution;
     std::vector<bool> neighborhood_flag;
-    std::set<int> neighbors;
-    int contribution_sum;
+    std::set<long> neighbors;
+    long contribution_sum;
 };
 
 struct Sample {
     Graph graph;
-    std::vector<int> mapping;
+    std::vector<long> mapping;
 };
 
 /// Detaches island and 1-degree vertices from the graph.
 Sample detach(const Graph &graph);
 
 /// Adds `vertex` to the expansion snowball sample.
-void es_add_vertex(const Graph &graph, ES_State &state, std::vector<int> &sampled, std::vector<int> &mapping,
-                   int vertex);
+void es_add_vertex(const Graph &graph, ES_State &state, std::vector<long> &sampled, std::vector<long> &mapping,
+                   long vertex);
 
 /// Updates the contribution of `vertex`, which has just been placed in the neighborhood of the current sample.
 /// Also decreases the contribution of all vertices that link to `vertex`.
-void es_update_contribution(const Graph &graph, ES_State &state, const std::vector<int> &mapping, int vertex);
+void es_update_contribution(const Graph &graph, ES_State &state, const std::vector<long> &mapping, long vertex);
 
 /// Samples vertices using the expansion snowball algorithm of Maiya et al.
 Sample expansion_snowball(const Graph &graph);
@@ -47,7 +47,7 @@ Sample expansion_snowball(const Graph &graph);
 Blockmodel extend(const Graph &graph, const Blockmodel &sample_blockmodel, const Sample &sample);
 
 /// Creates a Sample from sampled vertices and their mappings.
-Sample from_vertices(const Graph &graph, const std::vector<int> &vertices, const std::vector<int> &mapping);
+Sample from_vertices(const Graph &graph, const std::vector<long> &vertices, const std::vector<long> &mapping);
 
 /// Samples vertices with the highest degrees.
 Sample max_degree(const Graph &graph);
